refactor: get rid of Twap order statuses UI duplications
